modifications to the existing applications

+++ src/main/java/com/assignment/payslip/tax/TaxCalculation.java
@@ -1,6 +1,8 @@
 package com.assignment.payslip.tax;
 
+import java.math.BigDecimal;
+
 public interface TaxCalculation {
 	
-	float getTaxAmount(int salary);
+	BigDecimal getTaxAmount(int salary);
 }
